agent/requirements.txt
websockets==12.0
docker==7.1.0
python-dotenv==1.0.1
httpx==0.27.0
aiohttp==3.9.5
